+++ src/classes/managers/trader/errors/get-trader-deals.error.ts
@@ -1,0 +1,9 @@
+import { cTraderXError } from '../../../models/ctrader-x-error.model';
+
+export class GetTraderDealsError extends cTraderXError {
+    constructor(error: unknown) {
+        super(
+            `Get trader deals error: ${cTraderXError.getMessageError(error)}`,
+        );
+    }
+}