+++ src/classes/managers/trader/trader.manager.ts
@@ -5,6 +5,11 @@ import { ICredentials } from '../models/credentials.model';
 import { GetTraderError } from './errors/get-trader.error';
 import { ProtoOATraderRes } from './proto/messages/ProtoOATraderRes';
 import { ProtoOATraderReq } from './proto/messages/ProtoOATraderReq';
+import { ProtoOADealListRes } from './proto/messages/ProtoOADealListRes';
+import { ProtoOADealListReq } from './proto/messages/ProtoOADealListReq';
+import { GetTraderDealsError } from './errors/get-trader-deals.error';
+
+type BaseProto = 'payloadType' | 'ctidTraderAccountId';
 
 export class TraderManager extends BaseManager {
     constructor(
@@ -37,6 +42,35 @@ export class TraderManager extends BaseManager {
         }
 
         this.logCallAttemptSuccess(this.getTrader);
+
+        return result;
+    }
+
+    async getTraderDeals(
+        req: Omit<ProtoOADealListReq, BaseProto>,
+    ): Promise<ProtoOADealListRes> {
+        this.logCallAttempt(this.getTraderDeals);
+
+        const payload: ProtoOADealListReq = {
+            ...req,
+            ctidTraderAccountId: this.credentials.ctidTraderAccountId,
+        };
+
+        let result: ProtoOADealListRes;
+        try {
+            result = (await this.connection.sendCommand(
+                ProtoOADealListReq.name,
+                payload,
+            )) as ProtoOADealListRes;
+        } catch (e) {
+            throw this.handleCTraderCallError(
+                e,
+                this.getTraderDeals,
+                new GetTraderDealsError(e),
+            );
+        }
+
+        this.logCallAttemptSuccess(this.getTraderDeals);
 
         return result;
     }